fix: api 요청 방식 next api 활용하도록 변경
